requirements.txt
langchain_huggingface
langchain
flask 
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_groq
langchain_experimental
-e .

